compiler: extract program to separate file
Ast should not contain implementation for program. We can use it
without Program.
